requirements.txt
# ═══════════════════════════════════════
# Backup Manager v2.2.9 — Dependencies
# Python 3.13+ required
# ═══════════════════════════════════════

# ── System Tray Icon ──
pystray>=0.19.5                # Notification area icon (shield, status colors)
Pillow>=10.0.0                 # Icon generation (PIL) + window icon
plyer>=2.1.0                   # Native Windows toast notifications (optional fallback)

# ── Encryption ──
cryptography>=42.0.0           # AES-256-GCM encryption with PBKDF2 key derivation

# ── Cloud Storage: Amazon S3 + S3-compatible ──
boto3>=1.34.0                  # AWS S3, MinIO, Wasabi, OVH, Scaleway, Backblaze B2, Cloudflare R2

# ── Cloud Storage: Azure ──
azure-storage-blob>=12.19.0   # Azure Blob Storage

# ── Cloud Storage: Google Cloud ──
google-cloud-storage>=2.14.0  # Google Cloud Storage

# ── SFTP ──
paramiko>=3.4.0                # SSH/SFTP transfers (password or key auth)

# ── Notes ──
# All dependencies are OPTIONAL — the app starts without them.
# Missing features are grayed out in the UI.
# On first launch, the installer auto-installs pystray, Pillow, plyer, and cryptography.
# Cloud/SFTP packages are installed only if the user selects those storage types.
# Proton Drive uses rclone (external binary, not a pip package).
